todo ISSUE

+++ lib/core/visualizer_frame.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 import 'constants.dart';
 
 class VisualizerFrame {
-  final Float64List magnitudes;
+  final Float32List magnitudes;
 
   VisualizerFrame({required this.magnitudes})
     : assert(magnitudes.length == kFftSize ~/ 2);
